Extract common methods to a common util class.
